requirements.txt
asgiref==3.7.1
Django==3.2.7
Pillow==9.0.0
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.6.2

djangorestframework~=3.14.0